+++ jdaco_dev/archived_eval_code/PPIComp_hemato/build_diff_networks.java
@@ -56,7 +56,7 @@ public class build_diff_networks {
 			Utilities.writeEntries(all_alt_splice_switches.keySet(), out_path + state1 + "_" + state2 + "_contributing_AS_proteins.txt");
 			
 			
-			double P_rew_rounded = (double) Math.round( rd1.getP_rew() * 1000d) / 1000d;
+			double P_rew_rounded = Math.round( rd1.getP_rew() * 1000d) / 1000d;
 			System.out.println(rd1.getNumberOfComparisons()  + " comparisons, " + "P_rew: " + P_rew_rounded + ", " + rd1.getInteractionReasonsCountMap().size() + " dIAs" );
 			
 			System.out.println(major_alt_splice_switches.keySet().size() + " alt. spliced proteins are the major reason that affect " + Utilities.getValueSetFromMultimap(major_alt_splice_switches).size() + " diff. interactions.");
@@ -85,7 +85,7 @@ public class build_diff_networks {
 				Utilities.writeEntries(all_alt_splice_switches2.keySet(), out_path + state1 + "_" + state3 + "_contributing_AS_proteins.txt");
 				
 				
-				double P_rew_rounded2 = (double) Math.round( rd2.getP_rew() * 1000d) / 1000d;
+				double P_rew_rounded2 = Math.round( rd2.getP_rew() * 1000d) / 1000d;
 				System.out.println(rd2.getNumberOfComparisons()  + " comparisons, " + "P_rew: " + P_rew_rounded2 + ", " + rd2.getInteractionReasonsCountMap().size() + " dIAs" );
 				
 				System.out.println(major_alt_splice_switches2.keySet().size() + " alt. spliced proteins are the major reason that affect " + Utilities.getValueSetFromMultimap(major_alt_splice_switches2).size() + " diff. interactions.");
